[refactor] Moved splat management UI to separate SplatManager component for better UI organization; Added support for gallery splat name retrieval
